make the cache work better
now with autoguessing of the mimetype, errors, taking advantage of activestorage and more
