+++ Gemfile
@@ -92,6 +92,7 @@ gem "addressable"
 # gem "loofah"
 # gem "ruby-readability"
 # gem "stopwords-filter"
+gem "mimemagic"
 
 # Utils
 gem "mustermann" # Used by the autumn moon bot framework for pattern matching
